Enforcing float32 for all fit inputs to prevent numerical precision errors. Moved linelist tables to float32. Fixed q3dpro [OI]/Ha error. PG1411 cube now fits through.

+++ q3dfit/fitloop.py
@@ -155,7 +155,8 @@ def fitloop(ispax, colarr, rowarr, cube, q3di, listlines, specConv,
                     for line in q3di.lines:
                         listlinesz[line] = \
                             np.array(listlines['lines']
-                                     [(listlines['name'] == line)]) * \
+                                     [(listlines['name'] == line)],
+                                     dtype='float32') * \
                                 (1. + q3di.zinit_gas[line][i, j, ])
 
             if q3di.docontfit:
